Move verifiers into their own namespace
